Update allocator to use API and merge into one crate
